+++ Packages/com.texelsaur.video/Editor/Inspectors/Component/URL Source/PlaylistInspector.cs
@@ -11,6 +11,7 @@ namespace Texel
         protected SerializedProperty trackCatalogModeProperty;
         protected SerializedProperty immediateProperty;
         protected SerializedProperty resumeAfterLoadProperty;
+        protected SerializedProperty interruptibleProperty;
 
         protected SerializedProperty playlistCatalogProperty;
         protected SerializedProperty playlistDataProperty;
@@ -28,6 +29,7 @@ namespace Texel
             trackCatalogModeProperty = serializedObject.FindProperty(nameof(Playlist.trackCatalogMode));
             immediateProperty = serializedObject.FindProperty(nameof(Playlist.immediate));
             resumeAfterLoadProperty = serializedObject.FindProperty(nameof(Playlist.resumeAfterLoad));
+            interruptibleProperty = serializedObject.FindProperty(nameof(Playlist.interruptible));
             playlistCatalogProperty = serializedObject.FindProperty(nameof(Playlist.playlistCatalog));
             playlistDataProperty = serializedObject.FindProperty(nameof(Playlist.playlistData));
             queueProperty = serializedObject.FindProperty(nameof(Playlist.queue));
@@ -50,6 +52,7 @@ namespace Texel
             EditorGUILayout.PropertyField(trackCatalogModeProperty);
             EditorGUILayout.PropertyField(immediateProperty);
             EditorGUILayout.PropertyField(resumeAfterLoadProperty);
+            EditorGUILayout.PropertyField(interruptibleProperty);
 
             EditorGUILayout.Space();
             EditorGUILayout.LabelField("Data", EditorStyles.boldLabel);